renamed actions to actionTypes and removed React Component from import
