+++ src/App.jsx
@@ -1,15 +1,13 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { Provider } from 'react-redux';
-import styled from 'styled-components';
-
-import store from './store';
-
 import MainContainer from './containers/MainContainer.jsx';
 
 const App = () => (
-  <Provider>
+  // <Provider>
+  <div>
     <MainContainer />
-  </Provider>
+  </div>
+  // </Provider>
 );
 
 export default App;
